feat: let users register
add endpoints for users to register. The only thing they need is a webId. Additionally, they can provide 'person' and 'address' information which will be used to generate an example VC for BRP and BRK. When this information is not provided, generated some fake data, so that they can still use the demonstrator